[V2] code refactoring and simon profile
1. maven version 충돌 일원화
2. 코드 리팩토링
3. simon stopwatch 추가로 프로듀셔, 컨슈머 프로파일값 계산 로직값 추가
4. 로그 레벨 정리

+++ src/main/java/com/claud/kafka/producer/vo/log/WithdrawLog.java
@@ -6,7 +6,7 @@ import com.claud.kafka.producer.vo.send.LogKey;
 public class WithdrawLog extends BaseVo {
     private int userNumber;
     private String accountNumber;
-    private long outputMoney;
+    private int outputMoney;
 
     public WithdrawLog(int userNumber, String accountNumber, int outputMoney) {
         super(LogType.DEPOSIT_LOG);
@@ -31,11 +31,11 @@ public class WithdrawLog extends BaseVo {
         this.accountNumber = accountNumber;
     }
 
-    public long getOutputMoney() {
+    public int getOutputMoney() {
         return outputMoney;
     }
 
-    public void setOutputMoney(long outputMoney) {
+    public void setOutputMoney(int outputMoney) {
         this.outputMoney = outputMoney;
     }
